refactor: restructure AGC with separate modules for settings, helpers, and calculations
- Extract presets and config struct into `config.rs`
- Split helper functions into new `helpers.rs` file
- Move `RMS` calculation to `rms.rs` module
- Move slowdown state to `slowdown.rs` module
